Add files via upload
This project is based on Intel - RealSense, and is totally based on face tracking. This game (app) is being designed for big food-cafe and restaurants for those foodie people who wants to ply with food. This Video Link above will show a small glimpse of how to play and how Intel realSense track face-mesh